ENG-1319 finished implementation of Install endpoint with conflict management

+++ src/main/java/org/entando/kubernetes/model/bundle/installable/AssetInstallable.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.util.concurrent.CompletableFuture;
 import lombok.extern.slf4j.Slf4j;
 import org.entando.kubernetes.client.core.EntandoCoreClient;
+import org.entando.kubernetes.controller.digitalexchange.job.model.InstallRequest.InstallAction;
 import org.entando.kubernetes.model.bundle.ComponentType;
 import org.entando.kubernetes.model.bundle.descriptor.AssetDescriptor;
 
@@ -13,12 +14,12 @@ public class AssetInstallable extends Installable<AssetDescriptor> {
     private final File file;
     private final EntandoCoreClient engineService;
 
-    public AssetInstallable(EntandoCoreClient service, AssetDescriptor assetDescriptor) {
-        this(service, assetDescriptor, null);
+    public AssetInstallable(EntandoCoreClient service, AssetDescriptor assetDescriptor, InstallAction action) {
+        this(service, assetDescriptor, null, action);
     }
 
-    public AssetInstallable(EntandoCoreClient service, AssetDescriptor assetDescriptor, File file) {
-        super(assetDescriptor);
+    public AssetInstallable(EntandoCoreClient service, AssetDescriptor assetDescriptor, File file, InstallAction action) {
+        super(assetDescriptor, action);
         this.file = file;
         this.engineService = service;
     }
@@ -27,7 +28,15 @@ public class AssetInstallable extends Installable<AssetDescriptor> {
     public CompletableFuture<Void> install() {
         return CompletableFuture.runAsync(() -> {
             log.info("Registering CMS Asset {}", getName());
-            engineService.createAsset(representation, file);
+            if (shouldSkip()) {
+                return; //Do nothing
+            }
+
+            if (shouldCreate()) {
+                engineService.createAsset(representation, file);
+            } else {
+                engineService.updateAsset(representation, file);
+            }
         });
     }
 
@@ -35,7 +44,9 @@ public class AssetInstallable extends Installable<AssetDescriptor> {
     public CompletableFuture<Void> uninstall() {
         return CompletableFuture.runAsync(() -> {
             log.info("Removing Content {}", getName());
-            engineService.deleteAsset("cc=" + getName());
+            if (shouldCreate()) {
+                engineService.deleteAsset("cc=" + getName());
+            }
         });
     }
